% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEnhancer.R
\docType{data}
\name{expression}
\alias{expression}
\title{Expression data:}
\description{
rownames should be genes
}
\details{
colnames should be cell types

the gene name and cell types are both case-sensitive

they should be used consistently across other input data as well.
}
\keyword{expression}
