% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEnhancer.R
\name{findEnhancer}
\alias{findEnhancer}
\title{\code{findEnhancer} is the main function that identifies enhancers for a given gene}
\usage{
findEnhancer(gene, expression, regulation_signal, regulation_tfbs,
  region_gene_mapping, min_tfb_events = 2, coCRE_corr_cutoff = 0.5,
  coCRE_cutoff = 1e+05, singleton_cutoff = 20000, alphaVal = 1,
  scale.predictors = TRUE, family = "gaussian", nfoldxval = 10, ...)
}
\arguments{
\item{gene}{gene of interest eg. Runx1 (case sensitive)}

\item{expression}{data.frame/matrix of expression with cell types as cols and genes as rows}

\item{regulation_signal}{data.frame/matrix of enhancer signals (active histone marks/DNaseI-seq/ATAC-seq etc) with cell types as cols and putative enhancer regions as rows.}

\item{regulation_tfbs}{data.frame/matrix with TF ChIP-seq data as cols and putative enhancer regions as rows. This is a binary matrix with 0 = no binding and 1 = binding}

\item{region_gene_mapping}{data.frame with region to gene mapping. Please refer to data(region_gene_mapping) for the format.}

\item{min_tfb_events}{the minimum number of TFB events per site to be consider for community CRE calculation. default = 2}

\item{coCRE_corr_cutoff}{the cutoff above which two putative enhancer regions (CREs) are considered to be correlated. Default = 0.5}

\item{coCRE_cutoff}{the cutoff beyond which the regions will not be considered for coCRE calculation. Default = 10000 (100kB)}

\item{singleton_cutoff}{the cutoff for considering singleton CRE. Default = 20000 (20Kb)}

\item{alphaVal}{alpha value for glmnet. Read glmnet manual}

\item{scale.predictors}{(TRUE/FALSE). Default = TRUE}

\item{family}{response family. Default = gaussian}

\item{nfoldxval}{n-fold cross validation for lmin/l1se. Default = 10 (for leave out cross validation this is equal to total number of cols in expression data)}
}
\value{
list of regions and p-values
}
\description{
\code{findEnhancer} is the main function that identifies enhancers for a given gene
}
