% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEnhancer.R
\docType{data}
\name{region_gene_mapping}
\alias{region_gene_mapping}
\title{Dataframe that maps regions to genes:}
\description{
All the values are case-sensitive
}
\details{
Colnames are case sensitive and should be preserved as shown.

chr: chromosome id

start: start of region

end: end of region

peak: name of the region.

gene: name of gene.

distance: shortest dist between region and peak. If the peak is within a gene, you can use any value that is less than "singleton_cutoff"

colnames should be cell types

the gene name and cell types are both case-sensitive

they should be used consistently across other input data as well.
}
\keyword{metadata}
