% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEnhancer.R
\docType{data}
\name{regulation_signal}
\alias{regulation_signal}
\title{Regulation Signal}
\description{
DataFrame with row.names as regions and col.names as cell types
}
\details{
Region names should be preserved in region_gene_mapping and regulation_tfbs

Cell types names should be preserved in regulation_tfbs and expression DFs
}
