% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findEnhancer.R
\docType{data}
\name{regulation_tfbs}
\alias{regulation_tfbs}
\title{Regulation TFBS}
\description{
DataFrame with row.names as regions and col.names as TF ChIP-seq.
}
\details{
A binary data, with 0 denoting absence of binding and 1 denoting binding

Region names should be preserved in region_gene_mapping and regulation_signal

Cell types names should be preserved in regulation_signal and expression DFs
}
